# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "swagger first try",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "product"
                ],
                "summary": "list",
                "description": "list all products",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "responses": {
                    "200": {
                        "description": "all listed products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Electronics"
                                            },
                                            "image_path": {
                                                "type": "string",
                                                "example": "images/electronics.jpg"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "product"
                ],
                "summary": "Create a new product with multiple images",
                "operationId": "785644b7d7dd38af4b51dafd0fbc1b6d",
                "requestBody": {
                    "description": "Product details and images",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "price",
                                    "images[]"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "S25"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "RAM 12GB"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 1099.99
                                    },
                                    "images[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product created"
                                        },
                                        "product_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/products/{product}": {
            "get": {
                "tags": [
                    "product"
                ],
                "summary": "show",
                "operationId": "db00b523e69d5dbc09c6edad76374bdc",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "product_id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "show product detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "title": {
                                            "type": "string",
                                            "example": "Electronics"
                                        },
                                        "image_path": {
                                            "type": "string",
                                            "example": "images/electronics.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "product not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "product"
                ],
                "summary": "update a product",
                "operationId": "525bf6203301bf5a9fae0c6ca351cba0",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "product_id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Product details and images",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "price",
                                    "images[]"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "S25"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "RAM 12GB"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 1099.99
                                    },
                                    "images[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product updated"
                                        },
                                        "product_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "product"
                ],
                "summary": "Delete a category",
                "operationId": "ac878db81fdb1333b21fc79e574a5a17",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "product_id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "product deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Enter your Bearer token in the format: Bearer {token}",
                "bearerFormat": "Passport",
                "scheme": "bearer"
            }
        }
    },
...
}